[FIX] Fix proc.go SA9004: use uint64 for both constants
Previous fix used float64 which caused type mismatch with uint64
multiplication on line 45. Both constants are used in uint64 arithmetic,
so both must be uint64.

Verified locally with cross-compilation: GOOS=linux GOARCH=amd64
and golangci-lint with Linux target.

+++ connector/collector/proc.go
@@ -10,8 +10,8 @@ import (
 var sysMemTotal = getSysMemTotal()
 
 const (
-	clockTicksPerSecond  uint64  = 100
-	nanoSecondsPerSecond float64 = 1e9
+	clockTicksPerSecond  uint64 = 100
+	nanoSecondsPerSecond uint64 = 1e9
 )
 
 func getSysMemTotal() int64 {
